feat: adjusting project structure
