+++ tsconfig.json
@@ -1,6 +1,15 @@
 {
   "extends": "expo/tsconfig.base",
   "compilerOptions": {
-    "strict": true
+    "strict": true,
+    "baseUrl": "./",
+    "paths": {
+      "@assets/*": [ "./src/assets/*" ],
+      "@components/*": [ "./src/components/*" ],
+      "@screens/*": [ "./src/screens/*" ],
+      "@storage/*": [ "./src/storage/*" ],
+      "@theme/*": [ "./src/theme/*" ],
+      "@utils/*": [ "./src/utils/*" ],
+    }
   }
 }
